Pin dependencies in Dockerfile
Fixes #12.
syncabook specifies its direct dependencies in setup.py.
It could pin all the dependencies since it's a tool to be used by end-users
but I'm not sure if it's a good idea when installed as a global python library.
But we can surely pin everything when building a Docker image.